vector update1
