+++ Paula/practise01/src/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(PiGenerator PiGenerator.cpp)
+target_include_directories(PiGenerator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
